third_party/WebKit/LayoutTests/http/tests/security/mixedContent/laxChecking/websocket/insecure-websocket-in-secure-page-worker-expected.txt

CONSOLE WARNING: The page at 'https://127.0.0.1:8443/security/mixedContent/websocket/resources/expect-throw-on-construction-worker.html' was loaded over HTTPS, but ran insecure content from 'ws://127.0.0.1:8880/echo': this content should also be loaded over HTTPS.

CONSOLE WARNING: Connecting to a non-secure WebSocket server from a secure origin is deprecated.
This is a testharness.js-based test.
FAIL Test that an insecure WebSocket connection is blocked when created in a worker created from HTTPS resource. assert_equals: evt.data expected "DONE" but got "FAIL: No exception was thrown"
Harness: the test ran to completion.
